% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsor.R, R/init.R
\name{break_print_line}
\alias{break_print_line}
\alias{break_print_line}
\title{Break a long line for pretty printing}
\usage{
break_print_line(line)

break_print_line(line)
}
\description{
Break a long line for pretty printing

Break a long line for pretty printing
}
\keyword{internal}
