% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsor.R
\name{gso_getvar2}
\alias{gso_getvar2}
\title{Extract variables from GUESS smart output into a tibble}
\usage{
gso_getvar2(gso, varname, pft, timestart = 1, timeend = length(gso$time))
}
\arguments{
\item{gso}{a gso object as return from gso_open}

\item{varname}{name of Patch or Pft variable}

\item{pft}{Name of the selected pft}
}
\value{
a tibble
}
