% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsor.R
\name{gso_getvar_raw}
\alias{gso_getvar_raw}
\title{Extract variables from GUESS smart output into a tibble}
\usage{
gso_getvar_raw(gso, varname)
}
\arguments{
\item{gso}{a gso object as return from gso_open}

\item{varname}{name of Patch or Pft variable}
}
\value{
an array
}
