% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{gso_init}
\alias{gso_init}
\title{Open a GUESS smart output netcdf file}
\usage{
gso_init(filename, ...)
}
\arguments{
\item{filename}{}

\item{...}{additional parameters passed to \code{nc_open}}
}
\value{
an object of class gso
}
\description{
Open a GUESS smart output netcdf file
}
\details{
Open a GUESS smart output netcdf file
}
