% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsor.R, R/init.R
\name{gso_getvar}
\alias{gso_getvar}
\alias{gso_getvar}
\title{Extract variables from GUESS smart output into a tibble}
\usage{
gso_getvar(gso, vars, newnames = NULL, silent = FALSE)

gso_getvar(gso, vars, newnames = NULL, silent = FALSE)
}
\arguments{
\item{gso}{a gso object as return from gso_open}

\item{vars}{character vector with names of variable to extract}

\item{newnames}{optional character vector for renaming the variables}

\item{silent}{supress status updates}

\item{gso}{a gso object as return from gso_open}

\item{vars}{character vector with names of variable to extract}

\item{newnames}{optional character vector for renaming the variables}

\item{silent}{supress status updates}
}
\value{
a tibble

a tibble
}
\description{
Extract variables from GUESS smart output into a tibble

Extract variables from GUESS smart output into a tibble
}
\details{
.. content for \details{} ..

.. content for \details{} ..
}
